f tests f json_api settings f pytest.ini
