+++ example/views.py
@@ -23,8 +23,7 @@ from example.serializers import (
     EntryDRFSerializers,
     EntrySerializer,
     ProjectSerializer,
-    ProjectTypeSerializer,
-    AuthorWithNestedFieldsSerializer)
+    ProjectTypeSerializer)
 
 HTTP_422_UNPROCESSABLE_ENTITY = 422
 
@@ -173,11 +172,6 @@ class NoFiltersetEntryViewSet(EntryViewSet):
 class AuthorViewSet(ModelViewSet):
     queryset = Author.objects.all()
     serializer_class = AuthorSerializer
-
-
-class AuthorWithNestedFieldsViewSet(ModelViewSet):
-    queryset = Author.objects.all()
-    serializer_class = AuthorWithNestedFieldsSerializer
 
 
 class CommentViewSet(ModelViewSet):
