+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
 DJANGO_SETTINGS_MODULE=example.settings.test
 filterwarnings =
-    ignore::DeprecationWarning
-    ignore::PendingDeprecationWarning
+    error::DeprecationWarning
+    error::PendingDeprecationWarning
